siistitty portaalin koodia ja tyylitelty bid card
