+++ frontend/src/components/Portal/SinglePostPortalView.js
@@ -1,21 +1,14 @@
-import { Box, Button, Container, Typography } from '@mui/material'
+import { Box, Container, Typography } from '@mui/material'
 import React from 'react'
-import { useDispatch, useSelector } from 'react-redux'
+import { useSelector } from 'react-redux'
 import SingleFeedPostInfo from '../Feed/SingleFeedPostInfo'
-import { Link, useParams } from 'react-router-dom'
-import { useNotification } from '../../hooks'
-import { modifyBidApprovedState, removBidFromPortalpost } from '../../reducers/portalPosts'
+import { useParams } from 'react-router-dom'
 import Togglable from '../Togglable'
 import MakeBidForm from './MakeBidForm'
-import CheckCircleIcon from '@mui/icons-material/CheckCircle'
-import { removePortalBid, updatePortalBid } from '../../reducers/portalBids'
+import PortalBidCard from './PortalBidCard'
 
 const SinglePostPortalView = () => {
     const user = useSelector(({user}) => user)
-
-    const dispatch = useDispatch()
-  const notifyWith = useNotification()
-
 
     const {id} = useParams()
 
@@ -23,37 +16,7 @@ const SinglePostPortalView = () => {
     const portalBidsToShow = useSelector(({portalBids}) => portalBids)
 
   // Funktioita
-  const handleAcceptbid = async (bidId) => {
-    const confirmed = window.confirm('Haluatko varmasti hyväksyä tämän tarjouksen?')
-    if (!confirmed) {
-      return // If the user clicks "Cancel," do nothing
-    }
-    try {
-      // tää pitää muuttaa myös käyttään eri statee
-      //dispatch(modifyBidApprovedState(bidId, post.id))
-      dispatch(updatePortalBid(bidId))
-      notifyWith('Tarjous hyväksytty', 'success')
-    } catch (error) {
-      notifyWith('Tarjouksen hyväksyntä epäonnistui', 'error')
-    }
-
-  }
-
-  const handleDeletebid = async (bidId) => {
-    const confirmed = window.confirm('Haluatko varmasti poistaa tämän tarjouksen?')
-    if (!confirmed) {
-      return // If the user clicks "Cancel," do nothing
-    }
-    try {
-      dispatch(removePortalBid(bidId))
-      // tämä on vanha dispatch(removBidFromPortalpost(bidId, post.id))
-      notifyWith('Poistettu onnistuneesti', 'success')
-      //dispatch(initializePortalBids())
-    } catch (error) {
-      notifyWith('Tarjouksen poisto epäonnistui', 'error')
-    }
-
-  }
+  
 
   // portaalipostauksen voi nähdä sen lisännyt käyttäjä ja kehittäjät
     if (!user || !post || (user.userType === 'regular' && post.user.id !== user.id)) {
@@ -85,35 +48,8 @@ const SinglePostPortalView = () => {
     marginBottom: '1rem', borderBottom: '1px solid black', textAlign: 'center' }}>Tarjouksesi tähän ilmoitukseen</Typography>
         <Box>
           {user && developerBidsOnPost.length > 0 ? developerBidsOnPost.map(offer => (
-            <Box key={offer.id}
-            sx={{
-              padding: '1rem',
-              backgroundColor: '#f0f0f0',
-              borderRadius: '0.5rem',
-              textDecoration: 'none',
-              color: 'black',
-              marginLeft: '3rem',
-              marginRight: '3rem',
-              display: 'flex',
-              marginTop: '1rem',
-              transition: '0.3s ease',
-              flexDirection: 'column',
-              '@media (max-width: 820px)': {
-                marginLeft: '0.1rem',
-                marginRight: '0.1rem',
-              },
-            }}
-            >
-              {offer.isApproved && (
-                <Typography>Tarjous hyväksytty <CheckCircleIcon/></Typography>
-              )}
-              <Typography>Hinta: {offer.price} euroa</Typography>
-              <Typography><Button component={Link} to={`/kehittajat/${offer.user}`}>{offer.offeror}</Button></Typography>
-              <Typography>{offer.timeStamp.split('T')[0]}</Typography>
-              <Typography sx={{ whiteSpace: 'break-spaces' }}>{offer.description}</Typography>
-              {user && (user.id === post.user.id || user.id === offer.user.id) && (
-                <Button sx={{ color: 'red' }} onClick={() => handleDeletebid(offer.id)}>Poista tarjous</Button>
-              )}
+            <Box key={offer.id}>
+              <PortalBidCard offer={offer} post={post}/>
             </Box>
           )): <Typography sx={{ textAlign: 'center' }}>Et ole tarjonnut tähän vielä</Typography>}
         </Box>
@@ -125,44 +61,14 @@ const SinglePostPortalView = () => {
           <Typography sx={{ fontSize: '1.5rem', marginTop: '1rem',
     marginBottom: '1rem', borderBottom: '1px solid black', textAlign: 'center' }}>Tarjoukset</Typography>
       <Box>
-        {portalBidsToShow && portalBidsToShow.map(offer => (
-          <Box key={offer.id}
-          sx={{
-            padding: '1rem',
-            backgroundColor: '#f0f0f0',
-            borderRadius: '0.5rem',
-            textDecoration: 'none',
-            color: 'black',
-            marginLeft: '3rem',
-            marginRight: '3rem',
-            marginTop: '1rem',
-            display: 'flex',
-            transition: '0.3s ease',
-            flexDirection: 'column',
-            '@media (max-width: 820px)': {
-              marginLeft: '0.1rem',
-              marginRight: '0.1rem',
-            },
-          }}
-          >
-            {offer.isApproved && (
-              <Typography>Tarjous hyväksytty <CheckCircleIcon/></Typography>
-            )}
-            <Typography>Hinta: {offer.price} euroa</Typography>
-            <Typography><Button component={Link} to={`/kehittajat/${offer.user}`}>{offer.offeror}</Button></Typography>
-            <Typography>{offer.timeStamp.split('T')[0] || 'ei tietoa'}</Typography>
-            <Typography sx={{ whiteSpace: 'break-spaces' }}>{offer.description}</Typography>
-            {user && user.id === post.user.id && !offer.isApproved ? (
-              <Button onClick={() => handleAcceptbid(offer.id)}>Hyväksy tarjous</Button>
-            ): null}
-            {user && user.id === post.user.id && offer.isApproved ? (
-              <Button onClick={() => handleAcceptbid(offer.id)}>Epähyväksy tarjous</Button>
-            ): null}
-            {user && (user.id === post.user.id || user.id === offer.user) && (
-              <Button sx={{ color: 'red' }} onClick={() => handleDeletebid(offer.id)}>Poista tarjous</Button>
-            )}
+        {portalBidsToShow && portalBidsToShow.length > 0 ? (
+          portalBidsToShow.map(offer => (
+          <Box key={offer.id}>
+            <PortalBidCard offer={offer} post={post}/>
           </Box>
-        ))}
+        ))): (
+          <Typography sx={{ textAlign: 'center' }}>Ei vielä tarjouksia</Typography>
+        )}
       </Box>
           </Box>
       )}
